allocation size on GPU uses size_t to avoid overflow

+++ include/gpu/Particles_gpu.cuh
@@ -89,7 +89,7 @@ void particle_deallocate_host(struct particles*);
  */
 void particles_positions_alloc_device(
     struct particles_positions_gpu* part_pos,
-    struct particles_positions_gpu** part_pos_ptr, int length);
+    struct particles_positions_gpu** part_pos_ptr, size_t length);
 
 /**
  * Dealloc memory on gpu for particle positions
